[FEAT] 비행 기록 API 구축 #2

+++ src/main/java/io/rapa/shooongbackend/flight/dto/FlightRecordVo.java
@@ -6,7 +6,7 @@ import io.rapa.shooongbackend.flight.dto.postion.RotationRequest;
 public record FlightRecordVo(
     Long sequence,
     Long timestampUnixMs,
-    Long elapsedTimeSeconds,
+    Double elapsedTimeSeconds,
     PositionRequest position,
     RotationRequest rotation
 ) {
